+++ src/timessquare/worker/functions/__init__.py
@@ -3,6 +3,7 @@ from .create_check_run import create_check_run
 from .create_rerequested_check_run import create_rerequested_check_run
 from .ping import ping
 from .pull_request_sync import pull_request_sync
+from .replace_nbhtml import replace_nbhtml
 from .repo_added import repo_added
 from .repo_push import repo_push
 from .repo_removed import repo_removed
@@ -16,4 +17,5 @@ __all__ = [
     "compute_check_run",
     "create_check_run",
     "create_rerequested_check_run",
+    "replace_nbhtml",
 ]