+++ src/timessquare/worker/servicefactory.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import httpx
+from safir.arq import ArqMode, ArqQueue, MockArqQueue, RedisArqQueue
 from safir.github import GitHubAppClientFactory
 from safir.slack.blockkit import SlackException
 from sqlalchemy.ext.asyncio import async_scoped_session
@@ -10,8 +11,8 @@ from structlog.stdlib import BoundLogger
 
 from timessquare.config import config
 from timessquare.dependencies.redis import redis_dependency
+from timessquare.services.backgroundpage import BackgroundPageService
 from timessquare.services.githubrepo import GitHubRepoService
-from timessquare.services.page import PageService
 from timessquare.storage.nbhtmlcache import NbHtmlCacheStore
 from timessquare.storage.noteburstjobstore import NoteburstJobStore
 from timessquare.storage.page import PageStore
@@ -56,14 +57,30 @@ async def create_page_service(
     http_client: httpx.AsyncClient,
     logger: BoundLogger,
     db_session: async_scoped_session,
-) -> PageService:
+) -> BackgroundPageService:
     """Create a PageService for arq tasks."""
     redis = await redis_dependency()
+    arq_queue = await create_arq_queue()
 
-    return PageService(
+    return BackgroundPageService(
         page_store=PageStore(db_session),
         html_cache=NbHtmlCacheStore(redis),
         job_store=NoteburstJobStore(redis),
         http_client=http_client,
         logger=logger,
+        arq_queue=arq_queue,
     )
+
+
+async def create_arq_queue() -> ArqQueue:
+    """Create an ArqQueue for arq tasks."""
+    mode = config.arq_mode
+    if mode == ArqMode.production:
+        if not config.arq_redis_settings:
+            raise RuntimeError(
+                "The redis_settings argument must be set for arq in "
+                "production."
+            )
+        return await RedisArqQueue.initialize(config.arq_redis_settings)
+    else:
+        return MockArqQueue()